changed uintptr_t to size_t

+++ include/almalloc.h
@@ -5,18 +5,18 @@
 
 void* aligned_malloc(int aligment_bytes, long size)
 {
-  uint8_t* ptr;
-  uint8_t* aligned_ptr;
+  uintptr_t* ptr;
+  uintptr_t* aligned_ptr;
   int ptr_size;
 
-  ptr_size = sizeof(uint64_t);
+  ptr_size = sizeof(uintptr_t);
 
   ptr = malloc(aligment_bytes + size + ptr_size);
   aligned_ptr = ptr;
   aligned_ptr = aligned_ptr + ptr_size;
-  aligned_ptr += (uint64_t)aligned_ptr % (uint64_t)aligment_bytes;
+  aligned_ptr += (uintptr_t)aligned_ptr % (uintptr_t)aligment_bytes;
 
-  *((uint64_t*)aligned_ptr - 1) = (uint64_t)ptr;
+  *((uintptr_t*)aligned_ptr - 1) = (uintptr_t)ptr;
 
   return aligned_ptr;
 
@@ -24,11 +24,11 @@ void* aligned_malloc(int aligment_bytes, long size)
 
 void aligned_free(void* aligned_memory)
 {
-  uint64_t* ptr;
+  uintptr_t* ptr;
 
-  ptr = ((uint64_t*)aligned_memory - 1);
+  ptr = ((uintptr_t*)aligned_memory - 1);
 
-  ptr = (uint64_t*) *ptr;
+  ptr = (uintptr_t*) *ptr;
 
   free(ptr);
